Interim checkin - Refactor to create CucumberITGenerator based on the parallelScheme used.
