+++ src/it/junit/issue_43-outline_runner/src/test/resources/features/feature2.feature
@@ -1,0 +1,18 @@
+Feature: Feature1
+
+  Scenario: Generate runners per scenario
+    Given I have a feature with 2 scenarios
+    When I generate runners per scenario
+    Then a runner should be created for each scenario
+
+  Scenario Outline: Generate runners per example in a scenario outline 
+    Given I have a scenario with 3 examples
+    When I generate runners per scenario
+    Then a runner should be created for <example> example
+    
+    Examples:
+    | example |
+    | foo      |
+    | bar      |
+    | baz      |
+
