+++ src/main/java/com/github/timm/cucumber/generate/ParallelScheme.java
@@ -1,0 +1,5 @@
+package com.github.timm.cucumber.generate;
+
+public enum ParallelScheme {
+    FEATURE, SCENARIO
+}
